fix session run duration column

+++ app/database/migrations/2024_09_05_120821_create_session_service_runs_table.php
@@ -23,7 +23,7 @@ return new class extends Migration {
             $table->unsignedInteger('passed')->nullable();
             $table->unsignedInteger('failed')->nullable();
             $table->unsignedInteger('skipped')->nullable();
-            $table->unsignedInteger('duration')->nullable();
+            $table->float('duration')->nullable();
             $table->text('result_log')->nullable();
             $table->text('run_log')->nullable();
